Added JPA, docker compose support, fixed testcases
