+++ docker/docker-compose.yaml
@@ -1,0 +1,9 @@
+services:
+  postgres:
+    image: 'postgres:16.0'
+    environment:
+      - 'POSTGRES_DB=boot-graalvm'
+      - 'POSTGRES_USER=postgres'
+      - 'POSTGRES_PASSWORD=s3cr3t'
+    ports:
+      - '5432:5432'
