+++ src/main/java/com/giri/boot/graalvm/GraalVmApplication.java
@@ -6,6 +6,9 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class GraalVmApplication {
     public static void main(String[] args) {
+        for (String arg: args) { // let's examine arguments passed
+            System.out.println("arg: " + arg);
+        }
         SpringApplication.run(GraalVmApplication.class, args);
     }
