+++ src/main/resources/application-local.yml
@@ -3,6 +3,16 @@ spring:
     activate:
       on-profile: local
 
+  docker:
+    compose:
+      file: docker/docker-compose.yaml
+
+  datasource:
+    driver-class-name: org.postgresql.Driver
+    url: jdbc:postgres://localhost/boot-graalvm
+    username: postgres
+    password: s3cr3t
+
 app:
   greeting:
     message: running in Local
