+++ src/test/groovy/com/giri/boot/graalvm/controller/HelloControllerIS.groovy
@@ -1,8 +1,10 @@
 package com.giri.boot.graalvm.controller
 
+import com.giri.boot.graalvm.config.TestContainersConfiguration
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
 import org.springframework.boot.test.context.SpringBootTest
+import org.springframework.context.annotation.Import
 import org.springframework.test.context.ActiveProfiles
 import org.springframework.test.web.servlet.MockMvc
 import spock.lang.Specification
@@ -17,6 +19,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
  * created Nov 18, 2023
  */
 @SpringBootTest
+@Import(TestContainersConfiguration.class)
 @AutoConfigureMockMvc
 @ActiveProfiles('test')
 class HelloControllerIS extends Specification {
